Separate toc item into grid and list subcomponents
